GUI Rough Draft 1
This is roughly what I want the Player HUD to look like. In the top left corner there would be text for health, stamina, and numerical displays for both. As seen in pGUI_2 I want the portrait to slowly change as bars of hp are depleted and and I want the bars to break and resemble falling rocks similar to a cave in because I think that would work thematically with the game. For boss fights I would want a similar HUD element in the bottom right of the screen to mirror the player and status effects like crippled would be displayed with icons beneath the stamina bar. As the player gets more upgrades I want their HP and ST bars to slowly expand across the screen so that eventually the players HUD resembles boss characters to signify a growth in power. I also made a rough crippled icon so that testing movement and the feeling of being crippled/not crippled was more easily conveyed since that's what you guys are focusing on right now.